Finished the ECHO user part

+++ os_lesson/echo/u_server/server.c
@@ -11,46 +11,75 @@
 #include <semaphore.h>
 
 #include "server.h"
+#include "hash.h"
 
 void *server_process(void *arg);
 pthread_mutex_t info_mutex;
 server_info cfg;
+int char_cnt;         //字符计数
+int client_cnt;       //客户端计数
+char gMode[10];       //默认全局配置
+Hash *hash_table;
 
 int main()
 {
     int server_sockfd;
     int server_len;
     struct sockaddr_un server_addr;
-    int ret;
+    int ret = -1;
     pthread_t thread_s;
     void *thread_result;
 
     memset(&cfg, 0, sizeof(cfg));
     memcpy(cfg.mode, UPPER, sizeof(UPPER));
+    memcpy(gMode, UPPER, sizeof(UPPER));
+    char_cnt = 0;
+    client_cnt = 0;
+
+	//创建哈希表
+	hash_table = hash_new();
+    if (hash_table == NULL) {
+        perror("server hash_new");
+
+        goto err;
+    }
 
     //创建
     unlink(SERVER_ADDR);
     server_sockfd = socket(AF_UNIX, SOCK_STREAM, 0);
+    if (server_sockfd < 0) {
+        perror("server socket");
+
+        goto err_hash;
+    }
 
     //命名
     server_addr.sun_family = AF_UNIX;
     strcpy(server_addr.sun_path, SERVER_ADDR);
     server_len = sizeof(server_addr);
-    bind(server_sockfd, (struct sockaddr *)&server_addr, server_len);
+    ret = bind(server_sockfd, (struct sockaddr *)&server_addr, server_len);
+    if (ret != 0 ) {
+        perror("server bind");
+        goto err_hash;
+    }
 
     //连接
-    listen(server_sockfd, 5);
+    ret = listen(server_sockfd, 5);
+    if (ret != 0) {
+        perror("server listen");
+        goto err_hash;
+    }
 
     //设置线程相关功能
     ret = pthread_mutex_init(&info_mutex, NULL);
     if(ret != 0) {
         perror("server pthread_mutex_init");
-        exit(1);
+        goto err_sock;
     }
     ret = pthread_create(&thread_s, NULL, server_process, &server_sockfd);
     if(ret != 0) {
         perror("server pthread_create");
-        exit(1);
+        goto err_mutex;
     }
 
     while(1) {
@@ -63,9 +92,9 @@ int main()
         if (strncmp(buf, SHOW_INFO, sizeof(SHOW_INFO)-1) == 0) {
             pthread_mutex_lock(&info_mutex);
             printf("\nServer info:\n");
-            printf("Global mode: %s\n", cfg.mode);
-            printf("Characters: %d\n", cfg.char_cnt);
-            printf("Clients: %d\n", cfg.client_cnt);
+            printf("Global mode: %s\n", gMode);
+            printf("Characters: %d\n", char_cnt);
+            printf("Clients: %d\n", client_cnt);
             pthread_mutex_unlock(&info_mutex);
         }
         else if(strncmp(buf, QUIT, sizeof(QUIT)-1) == 0) {
@@ -73,14 +102,28 @@ int main()
         }
         else if(strncmp(buf, MODE, sizeof(MODE)-1) == 0) {
             sscanf (buf, "mode %[0-9] %[a-z]", client, mode);
-            pthread_mutex_lock(&info_mutex);
-            cfg.client_id = atoi(client);
-            memcpy(cfg.mode, mode, strlen(mode));
-            pthread_mutex_unlock(&info_mutex);
+            if (client[0] == '0') {
+                pthread_mutex_lock(&info_mutex);
+                memcpy(gMode, mode, strlen(mode));
+                pthread_mutex_unlock(&info_mutex);
+            }
+            else {
+                pthread_mutex_lock(&info_mutex);
+                ret = hash_get(hash_table, client, &cfg);
+                pthread_mutex_unlock(&info_mutex);
+                if(ret == 0) {  // 在哈希表中没有
+                    printf("not found this client %s", client);
+                    continue;
+                }
+                else {//在哈希表中找到
+                    pthread_mutex_lock(&info_mutex);
+                    memcpy(cfg.mode, mode, strlen(mode));
+                    hash_set(&hash_table, client, cfg);    //更新到哈希表中
+                    pthread_mutex_unlock(&info_mutex);
+                }
+            }
         }
     }
-
-    close(server_sockfd);
 
     ret = pthread_cancel(thread_s);
     if (ret != 0) {
@@ -94,12 +137,13 @@ int main()
         exit(1);
     }
 
-    ret = pthread_mutex_destroy(&info_mutex);
-    if (ret != 0) {
-        perror("server pthread_mutex_destroy");
-        exit(1);
-    }
-
+err_mutex:
+    pthread_mutex_destroy(&info_mutex);
+err_sock:
+    close(server_sockfd);
+err_hash:
+    hash_delete(hash_table);
+err:
     return ret;
 }
 
@@ -112,6 +156,7 @@ void upper_process(int fd, char data[], int len)
 
     for( i = 0; i < len; i++) {
         data[i] = toupper(data[i]);
+        usleep(200*1000);
     }
     //printf("upper server to client on fd %d\n", fd);
 
@@ -124,8 +169,21 @@ void lower_process(int fd, char data[], int len)
 
     for( i = 0; i < len; i++) {
         data[i] = tolower(data[i]);
+        usleep(200*1000);
     }
     //printf("tolower server to client on fd %d\n", fd);
+
+    return;
+}
+
+void normal_process(int fd, char data[], int len)
+{
+    int i = 0;
+
+    for( i = 0; i < len; i++) {
+        usleep(200*1000);
+    }
+    //printf("normal server to client on fd %d\n", fd);
 
     return;
 }
@@ -139,6 +197,7 @@ void *server_process(void *arg)
     int client_len;
     int server_sockfd = *(int *)arg;
     struct sockaddr_un client_addr;
+	char key_tag[20]  = {0};
 
     ret = pthread_setcancelstate(PTHREAD_CANCEL_ENABLE, NULL);
     if (ret != 0) {
@@ -167,6 +226,7 @@ void *server_process(void *arg)
             (struct timeval *)0);
         if(ret < 1) {
             perror("server select");
+            FD_CLR(fd, &readfds);
             exit(1);
         }
 
@@ -180,34 +240,52 @@ void *server_process(void *arg)
                     sprintf(buf, "%d", client_sockfd);
                     write(client_sockfd, buf, sizeof(buf));
                     //printf("adding client on fd %d\n", client_sockfd);
+
                     pthread_mutex_lock(&info_mutex);
-                    cfg.client_cnt++;
+                    client_cnt++;
+                    cfg.client_id = client_sockfd;
+                    memcpy(cfg.mode, gMode, strlen(gMode));
+                    hash_set(&hash_table, buf, cfg);
                     pthread_mutex_unlock(&info_mutex);
                 }
                 else { //client端处理
                     ioctl(fd, FIONREAD, &nread);
                     if(nread == 0) {
+                        sprintf(key_tag, "%d", fd);
                         close(fd);
                         FD_CLR(fd, &readfds);
+                        hash_remove(&hash_table, key_tag);
+                        pthread_mutex_lock(&info_mutex);
+                        client_cnt--;
+                        pthread_mutex_unlock(&info_mutex);
                         //printf("removing client on fd %d\n", fd);
                     }
                     else {
-                        if (fd != 0) { //todo 增加队列动态识别
+                        sprintf(key_tag, "%d", fd);
+                        pthread_mutex_lock(&info_mutex);
+                        ret = hash_get(hash_table, key_tag, &cfg);
+                        pthread_mutex_unlock(&info_mutex);
+                        if(ret == 0) {// 在哈希表中没有
+                            continue;
+                        }
+                        else {//在哈希表中找到
                             nread = read(fd, buf, sizeof(buf));
-                            sleep(1);
-                            pthread_mutex_lock(&info_mutex);
-                            cfg.char_cnt += (strlen(buf) - 1);
-                            pthread_mutex_unlock(&info_mutex);
-                            if (strncmp(cfg.mode, UPPER, sizeof(UPPER)-1) == 0)
-                                upper_process(fd, buf, nread);
+                            //printf("cfg.mode %s\n", cfg.mode);
+                            //printf("cfg.client_id %d\n", cfg.client_id);
 
-                            if (strncmp(cfg.mode, LOWER, sizeof(LOWER)-1) == 0)
-                                lower_process(fd, buf, nread);
+                            pthread_mutex_lock(&info_mutex);
+                            char_cnt += (strlen(buf) - 1);
+                            if (strncmp(cfg.mode, UPPER, sizeof(UPPER)-1) == 0){
+                                upper_process(fd, buf, strlen(buf));
+                            }
+                            else if(strncmp(cfg.mode, LOWER, sizeof(LOWER)-1) == 0){
+                                lower_process(fd, buf, strlen(buf));
+                            } else {
+                                normal_process(fd, buf, strlen(buf));
+                            }
+                            pthread_mutex_unlock(&info_mutex);
 
                             write(fd, buf, sizeof(buf));
-                        }
-                        else {
-                            printf("nonsupport this client %d", fd);
                         }
                     }
                 }
